+++ src/main/java/by/vk/betting/api/analytic/service/counter/PerGameScoreCounter.java
@@ -6,6 +6,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
@@ -24,7 +25,7 @@ public class PerGameScoreCounter
     var games = analytics.size();
     var name = analytics.get(ZERO).teamName();
     var amount = scored == ZERO ? ZERO : (double) scored / games;
-    var metric = new Metric(name, amount);
+    var metric = new Metric(name, BigDecimal.valueOf(amount));
     LOGGER.debug("[METRICS] Per game scored metric for [{}] is [{}]", entry.getKey(), metric);
     return metric;
   }
